reorganize the code
